add entire MIT license text
The conditions under which a user receives a piece of software should be unambiguous and as easy to consume as possible. This, along with the right kind of license, helps to remove risk from users in the event of an intellectual property dispute (whether on the part of the author or another interested party).

In my opinion, the easiest way to achieve this is to include the entire text of the license conditions with the software as it is distributed.

I've taken the liberty of including the entire "MIT" license (as specified here http://opensource.org/licenses/MIT), with the copyright holder being named as Dominic Tarr.

Please review the content of the license before accepting this patch!